+++ src/components/index.ts
@@ -41,7 +41,7 @@ export * from "./ui/navigation/SidebarBase";
 export * from "./ui/overlays/DrawerBase";
 export * from "./ui/overlays/PopoverBase";
 export * from "./ui/overlays/SheetBase";
-export { default as ErrorMessage } from "./ui/ErrorMessage";
+export { default as ErrorMessage } from "./ui/shared/ErrorMessage";
 export * from "./ui/UniversalTooltipRenderer";
 
 export * from "./ui/navigation/ContextMenuBase";